requirements.txt
Django
Pillow
django-cleanup
django-allauth
django-htmx
